Moved the TokenManager into its own folder and extracted the StreamString class.
