+++ LOLBITS/LOLBITS/Controlling/Controller.cs
@@ -6,6 +6,7 @@ using System.Reflection;
 using System.Text;
 using System.Threading;
 using LOLBITS.Loading;
+using LOLBITS.TokenManagement;
 using Newtonsoft.Json;
 using BITS4 = BITSReference4_0;
