+++ LOLBITS/LOLBITS/LOLBITS.csproj
@@ -79,7 +79,8 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Controlling\Response.cs" />
     <Compile Include="SysCallManager.cs" />
-    <Compile Include="TokenManager.cs" />
+    <Compile Include="TokenManagement\StreamString.cs" />
+    <Compile Include="TokenManagement\TokenManager.cs" />
     <Compile Include="Utils.cs" />
   </ItemGroup>
   <ItemGroup>
