+++ LOLBITS/LOLBITS/SyscallManager.cs
@@ -13,13 +13,10 @@ namespace LOLBITS
 
         private readonly byte[] _shellCode = {
                     0x4C, 0x8B, 0xD1,             // mov r10, rcx
-                    0xB8, 0x00, 0x00, 0x00, 0x00, // mov eax, 0x00 (syscall identifier)
-                    0x0F, 0x05,                   // syscall
+                    0xB8, 0x00, 0x00, 0x00, 0x00, // mov eax, 0x00 (sysCall identifier)
+                    0x0F, 0x05,                   // sysCall
                     0xC3                          // ret
             };
-
-
-
 
         public SysCallManager()
         {
