+++ LOLBITS/LOLBITS/Utils.cs
@@ -5,6 +5,7 @@ using System.Runtime.InteropServices;
 using System.Security.Principal;
 using System.Text;
 using System.Threading;
+using LOLBITS.TokenManagement;
 
 namespace LOLBITS
 {
